requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
requests==2.31.0